% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateAS.R
\name{annotateAS}
\alias{annotateAS}
\title{Annotate alternative-spliced segments from a GTF annotation}
\usage{
annotateAS(x, as.data.frame = FALSE, append = FALSE)
}
\arguments{
\item{x}{GRanges object containing transcript features in GTF format}

\item{as.data.frame}{If TRUE, function will return a dataframe instead of GRanges GTF object
(default = FALSE)}

\item{append}{If TRUE, function will append the alternative-spliced segments to x and return
a new GTF GRanges object}
}
\value{
GRanges object or data-frame containing alternative-spliced segments found in x
}
\description{
Annotate alternative-spliced segments from a GTF annotation
}
\author{
Fursham Hamid
}
