% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCDS.R
\name{buildCDS}
\alias{buildCDS}
\title{Construct query CDS using reference as guide}
\usage{
buildCDS(query, ref, fasta)
}
\arguments{
\item{query}{}

\item{ref}{}

\item{fasta}{BSgenome or Biostrings object containing genomic sequence}
}
\value{
GRanges object containing exon
}
\description{
buildCDS will attempt to construct query CDS information by
firstly deriving its start_codon. To do this, the program will:
(1) search transcript for an annotated start codon
(2) if above fails, search for an internal ATG codon
(3) if above fails, align the frame of query to reference
}
\author{
Fursham Hamid
}
