% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCovs.R
\name{calcCovs}
\alias{calcCovs}
\title{Obtain percent coverage between query and reference transcripts}
\usage{
calcCovs(query, ref, query2ref, ids = c(1, 2), return = c("best", "all"))
}
\arguments{
\item{query}{GRanges or GRangesList object containing exons for each query transcript. 
If query is a GRangesList object with length > 1, transcript names
have to be listed in query2ref dataframe}

\item{ref}{If query is a GRanges object, ref is a GRanges object containing CDS for a 
reference transcript. If query is a GRangesList object, ref is a GRangesList object 
containing CDS for each reference transcript. Ref transcript names have to be 
listed in query2ref dataframe}

\item{query2ref}{If query and ref are GRangesList objects, dataframe with at least 2 columns: 
query transcript_id and its reference transcript_id. Query and ref transcript_ids 
have to match transcript names in query and refCDS objects. Transcripts with 
missing corrresponding GRanges object will return error.
query2ref is not mandatory if query and ref are GRanges object or a GRangesList 
object of length 1}

\item{ids}{Numeric vector stating which columns of query2ref dataframe contain the
query and reference transcript_ids respectively.}

\item{return}{If query2ref contain query transcripts with multiple comparisons, function
will return comparisons with 'best' coverage value by default. Alternatively,
a full report can be returned by setting argument to 'all'}
}
\value{
Dataframe from query2ref with coverage values appended as a new column
}
\description{
Obtain percent coverage between query and reference transcripts
}
\examples{
calcCovs(query_exons, ref_exons, q2r)
}
\author{
Fursham Hamid
}
