% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputs.R
\name{checkInputs}
\alias{checkInputs}
\title{Check input files for workflow}
\usage{
checkInputs(query, ref, fasta)
}
\arguments{
\item{query}{GRanges object containing imported query GTF data}

\item{ref}{GRanges object containing reference GTF data}

\item{fasta}{BSgenome or Biostrings object containing genomic sequence}
}
\description{
Check input files for workflow
}
