% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{domains.out}
\alias{domains.out}
\title{Example output of predictDomains()}
\format{
A data.frame with 14880 rows and 5 columns:
\describe{
  \item{element}{Predicted domains from newly-discovered mRNAs}
  ...
}
}
\usage{
domains.out
}
\description{
Output dataframe from predictDomains() function.
}
\keyword{datasets}
