% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportExonsCDStoGTF.R
\name{exportExonsCDStoGTF}
\alias{exportExonsCDStoGTF}
\title{Merge GRangesList exons and cds objects and export as GTF}
\usage{
exportExonsCDStoGTF(exons, cds, con)
}
\arguments{
\item{exons}{GRangesList object containing exons
for each transcript.}

\item{cds}{GRangesList object containing cds
for each transcript.}

\item{con}{Path to file for saving GTF}
}
\description{
Merges two GRangesList objects; one containing exon coordinates and one containing cds coordinates.
}
