% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoverages.R
\name{getCoverages}
\alias{getCoverages}
\title{Obtain percent coverage between query and reference transcripts}
\usage{
getCoverages(query, ref, query2ref, ids = c(1, 2), return = c("best", "all"))
}
\arguments{
\item{query}{GRangesList object containing exons for each query transcript. Transcripts
have to be listed in query2ref dataframe, else coverage values will not be
calculated}

\item{ref}{GRangesList object containing CDS for each reference transcript.}

\item{query2ref}{Dataframe with at least 2 columns: query transcript_id and its
reference transcript_id. Query and ref transcript_ids have to match transcript
names in query and refCDS objects. Transcripts with missing corrresponding
GRanges object will return error}

\item{ids}{Numeric vector stating which columns of query2ref dataframe contain the
query and reference transcript_ids respectively.}

\item{return}{If query2ref contain query transcripts with multiple comparisons, function
will return comparisons with 'best' coverage value by default. Alternatively,
a full report can be returned by setting argument to 'all'}
}
\value{
Dataframe from query2ref with coverage values appended as a new column
}
\description{
Obtain percent coverage between query and reference transcripts
}
\examples{
getCoverages(query_exons, ref_exons, q2r)
}
\author{
Fursham Hamid
}
