% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_consistentSeqLevels.R
\name{has_consistentSeqlevels}
\alias{has_consistentSeqlevels}
\title{Test consistency in seqlevels across objects}
\usage{
has_consistentSeqlevels(...)
}
\arguments{
\item{...}{Two or more objects with seqlevels information}
}
\value{
Logical value as to whether all objects have consistent seqlevels
}
\description{
Test consistency in seqlevels across objects
}
\author{
Fursham Hamid
}
