% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyNMDexons.R
\name{identifyNMDexons}
\alias{identifyNMDexons}
\title{Find NMD-causing exons from GTF objects}
\usage{
identifyNMDexons(x, fasta, NMD.result = NULL, ConsScores = "none")
}
\arguments{
\item{x}{GRanges object containing query GTF data with CDS information.}

\item{fasta}{BSgenome or Biostrings object containing genomic sequence}

\item{NMD.result}{Output dataframe from predictNMD() [Optional]. If not provided, or if 
data-frame do not contain necessary variables, identifyNMDexons will run
predictNMD() first.}

\item{ConsScores}{Character value of the annotation database to use to calculate mean exon 
conservation scores. Database can be a Bioconductor annotation package or
one of the available database from AnnotationHub. Run
\code{GenomicScores::availableGScores()} to list the available 
databases. A vector of connservation score for each exon will be appended
as a metadata column in the output GRanges object.
By default ("none"), exon conservation scores will not be calculated.}
}
\value{
GRanges object containing coordinates of identified NMD-causing exons, 
including metadata on its gene of origin, type of AS exon, type of NMD exon
and whether it is spliced within CDS
}
\description{
This function will identify alternative exons that triggers NMD when 
spliced or skipped.
}
\details{
Alternative splicing of regulated exons may introduce premature termination 
codon which are cleared by nonsense-mediated decay. There are two major class 
of such exons: Poison and ORF-maintaining exons. The former triggers NMD when
spliced in, hence its name, while the latter triggers NMD when removed from a 
transcript.

To identify and classify NMD-causing exons, identifyNMDexons first find
alternative segments between NMD-sensitive transcripts and pre-determined 
reference transcript of the same gene. The longest, NMD-insensitive coding 
transcript of each gene is chosen as the best reference. For each alternative segment, 
a modified version of reference RNA is built by inserting exons spliced in NMD transcripts, or
by removing exons skipped from NMD transcripts. These modified transcripts are
then tested for NMD features and if positive, the alternative exon will be
annotated as NMD-causing.
}
\examples{
## ---------------------------------------------------------------------
## EXAMPLE USING SAMPLE DATASET
## ---------------------------------------------------------------------
library(BSgenome.Mmusculus.UCSC.mm10)

# build CDS information first
out.gtf <- buildCDS(matched_query_gtf, ref_gtf, Mmusculus)

# proceed with identifying NMD exons
NMD.exons_1 <- identifyNMDexons(out.gtf, Mmusculus)

# if output of predictNMD() has been generated, this can be provided to
# NMD.result argument
NMD.output <- predictNMD(out.gtf)
NMD.exons_2 <- identifyNMDexons(out.gtf, Mmusculus, NMD.result = NMD.output)

identical(NMD.exons_1, NMD.exons_2)


## ---------------------------------------------------------------------
## EXAMPLE USING TRANSCRIPT ANNOTATION
## ---------------------------------------------------------------------
\donttest{
library(AnnotationHub)

## Retrieve GRCm38 trancript annotation
ah <- AnnotationHub()
GRCm38_gtf <- ah[["AH60127"]]

## Identify NMD-causing exons in entire transcriptome
NMD.exons.all <- identifyNMDexons(GRCm38_gtf, Mmusculus)
}

}
