% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importFASTA}
\alias{importFASTA}
\title{Import FASTA file into R}
\usage{
importFASTA(con)
}
\arguments{
\item{con}{Path to FASTA file}
}
\value{
Imported DNAStringSet object
}
\description{
This function is a wrapper to Biostrings::readDNAStringSet() 
function to import FASTA genome sequence file and simultaneously 
convert long chromosome names 
(e.g. 1 dna:chromosome chromosome:GRCm38:1:1:195471971:1 REF) 
to short names (e.g. 1)
}
\author{
Fursham Hamid
}
