% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_gtf.R
\name{is_gtf}
\alias{is_gtf}
\title{Test for GTF GRanges object}
\usage{
is_gtf(...)
}
\arguments{
\item{...}{One or more objects to test}
}
\value{
Logical values as to whether each object is a GTF GRanges
}
\description{
Test for GTF GRanges object
}
\author{
Fursham Hamid
}
