% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicingFunctions.R
\name{labelSplicedSegment}
\alias{labelSplicedSegment}
\title{Label alternative-spliced segments from a GTF annotation}
\usage{
labelSplicedSegment(
  x,
  as.data.frame = FALSE,
  append = FALSE,
  group.by = "gene_id"
)
}
\arguments{
\item{x}{GRanges object containing transcript features in GTF format}

\item{as.data.frame}{If TRUE, function will return a dataframe instead of GRanges GTF object
(default = FALSE)}

\item{append}{If TRUE, function will append the alternative-spliced segments to x and return
a new GTF GRanges object (default = TRUE)}
}
\value{
GRanges object or data-frame containing alternative-spliced segments found in x
}
\description{
Label alternative-spliced segments from a GTF annotation
}
\examples{

## ---------------------------------------------------------------------
## EXAMPLE USING TOY DATASET
## ---------------------------------------------------------------------
require(GenomicRanges)

## Create toy GRanges GTF object
gr <- GRanges("chr1", IRanges(start = c(1, 101, 1, 51, 101), width = c(20, 20, 20, 10, 20)), "+",
  type = "exon", gene_id = "geneA",
  transcript_id = Rle(c("transcript1", "transcript2"), lengths = c(2, 3))
)

## Annotate alternative segments
labelSplicedSegment(gr)

## ---------------------------------------------------------------------
## EXAMPLE USING SAMPLE DATASET
## ---------------------------------------------------------------------
## Using GTF GRanges as input
labelSplicedSegment(query_gtf)

## Output as dataframe
labelSplicedSegment(query_gtf, as.data.frame = TRUE)

## Append AS info to input
labelSplicedSegment(query_gtf, append = TRUE)

## ---------------------------------------------------------------------
## EXAMPLE USING TRANSCRIPT ANNOTATION DATABASE
## ---------------------------------------------------------------------
\dontrun{
library(AnnotationHub)

## Retrieve GRCm38 trancript annotation
ah <- AnnotationHub()
GRCm38_gtf <- ah[["AH60127"]]

## Run tool on specific gene family
labelSplicedSegment(GRCm38_gtf)
}

}
\author{
Fursham Hamid
}
