% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGTF.R
\name{makeGTF}
\alias{makeGTF}
\title{Combine and export exons and cds Granges to GTF}
\usage{
makeGTF(exons, cds, con)
}
\arguments{
\item{exons}{GRangesList object containing exons for each query transcript.}

\item{cds}{GRangesList object containing cds for each query transcript.}

\item{con}{See ?rtracklayer::export for a description of this argument}
}
\value{
See ?rtracklayer::export for a description of this argument
}
\description{
Combine and export exons and cds Granges to GTF
}
