% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchChromosomes.R
\name{matchChromosomes}
\alias{matchChromosomes}
\title{Match seqlevels of input GRanges to reference GRanges or BioString objects}
\usage{
matchChromosomes(x, to)
}
\arguments{
\item{x}{GRanges object with seqnames to change}

\item{to}{GRanges object from which seqnames is referenced}
}
\value{
Corrected input GRanges
}
\description{
A convenient wrapper to match seqlevels of a query GRanges object to a reference
object that contain seqlevels information. Reference can be a GRanges, GRangesList,
BioString or DNAString object. Seqlevels which fail to match will be dropped.
}
\author{
Fursham Hamid
}
