% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matched_query_gtf}
\alias{matched_query_gtf}
\title{Seqlevels and gene_id matched query data}
\format{
A GRanges object with 56 ranges and 6 metadata columns:
\describe{
  \item{element}{Exon start and end coordinates from 4 transcripts}
  ...
}
}
\usage{
matched_query_gtf
}
\description{
query_gtf data which have been corrected for its seqlevels and gene_ids
}
\keyword{datasets}
