% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchSeqLevels.R
\name{matchSeqLevels}
\alias{matchSeqLevels}
\title{Match seqnames of input GRanges to reference GRanges}
\usage{
matchSeqLevels(from, to)
}
\arguments{
\item{from}{GRanges object with seqnames to change}

\item{to}{GRanges object from which seqnames is referenced}
}
\value{
Corrected input GRanges
}
\description{
Match seqnames of input GRanges to reference GRanges
}
\author{
Fursham Hamid
}
