% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeExonsCDS.R
\name{mergeExonsCDS}
\alias{mergeExonsCDS}
\title{Merge GRangesList exons and cds objects}
\usage{
mergeExonsCDS(exons, cds)
}
\arguments{
\item{exons}{GRangesList object containing exons
for each transcript.}

\item{cds}{GRangesList object containing cds
for each transcript.}
}
\value{
GRanges object
}
\description{
Merges two GRangesList objects; one containing exon coordinates and one containing cds coordinates.
}
