% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eachfunctions.R
\name{mutateeach}
\alias{mutateeach}
\title{Internally create or transform metadata of a GenomicRangesList}
\usage{
mutateeach(x, ...)
}
\arguments{
\item{x}{GRangesList object}

\item{...}{Name-value pairs of expressions. The name of each argument will be the name
of a new metadata column, and the value will be its corresponding value.}
}
\value{
Transformed GRangesList object
}
\description{
Internally create or transform metadata of a GenomicRangesList
}
\author{
Fursham Hamid
}
