% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{new_query_gtf}
\alias{new_query_gtf}
\title{Query data containing CDS information}
\format{
A GRanges object with 105 ranges and 7 metadata columns:
\describe{
  \item{ranges}{Chromosome, start, end, and strand info of 4 transcripts
  and its exons}
  \item{type}{Entry type; transcript or exon}
  \item{transcript_id}{ID given to transcripts}
  \item{gene_id}{Matched gene_id}
  \item{old_gene_id}{Original gene_id}
  \item{match_level}{Level of matching performed}
  \item{gene_name}{Name of gene}
  \item{phase}{Phase of open-reading frame}
  ...
}
}
\usage{
new_query_gtf
}
\description{
matched_query_gtf data that has undergone buildCDS function and containing
CDS features
}
\keyword{datasets}
