% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictNMD.R
\name{predictNMD}
\alias{predictNMD}
\title{Predict sensitivity of mRNA transcripts to NMD}
\usage{
predictNMD(x, cds = NULL, NMD_threshold = 50, which = NULL)
}
\arguments{
\item{x}{Can be a GRanges object containing exon and CDS transcript features in GTF
format.

Can be a GRangesList object containing exon features for a list of transcripts.

Can be a GRanges object containing exon features for a transcript.}

\item{cds}{If `x` is a GRangesList object, `cds` has to be a GRangesList containing CDS features
for the list of transcripts in `x`. List names in `x` and `cds` have to match.

If `x` is a GRanges object, `cds` has to be a GRanges containing CDS features
for the transcript in `x`.}

\item{NMD_threshold}{Minimum distance of stop_codon to last exon junction (EJ) which triggers NMD.
Default = 50bp}

\item{which}{List containing names of transcripts from `x` to subset for analysis}
}
\value{
Dataframe with prediction of NMD sensitivity and NMD features:

is_NMD: logical value in prediciting transcript sensitivity to NMD

dist_to_lastEJ: Integer value indicating distance of STOP codon to last EJ
Values are referenced from last EJ, thus a positive value indicates upstream
position of STOP codon while negative value indicates downstream position

num_of_down_EJs: Number of downstream EJs

dist_to_downEJs: Integer value indicating distance of STOP codon to each down
EJs. Values are referenced from last EJ, thus a positive value indicates
upstream position of STOP codon while negative value indicates downstream
position together with distances of STOP codon to last EJ
}
\description{
Predict sensitivity of mRNA transcripts to NMD
}
\examples{

### Examples with GRanges objects
predictNMD(query_exons$transcript1, query_cds$transcript1) # NMD-insensitive
predictNMD(query_exons$transcript3, query_cds$transcript3) # NMD-sensitive

### Examples with GRangesList object
predictNMD(query_exons, query_cds)
predictNMD(query_exons, query_cds, which = c("transcript1", "transcript3"))
}
\author{
Fursham Hamid
}
