% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{q2r}
\alias{q2r}
\title{Dataframe containing a list of matched query and reference IDs}
\format{A dataframe with 8 rows
\describe{
  \item{transcript_id}{List of transcript_id from query_gtf}
  \item{ref_transcript_id}{List of transcript_id from ref_gtf}
  ...
}}
\usage{
q2r
}
\description{
Dataframe with 2 columns, each containing a list of transcript_id from
query_gtf and ref_gtf dataset respectively
}
\keyword{datasets}
