% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{q2rcovs}
\alias{q2rcovs}
\title{Dataframe containing a list of matched query and reference IDs with coverage values}
\format{A dataframe with 8 rows
\describe{
  \item{transcript_id}{List of transcript_id from query_gtf}
  \item{ref_transcript_id}{List of transcript_id from ref_gtf}
  \item{coverage}{List of percent coverage values between query and ref transcripts}
  ...
}}
\usage{
q2rcovs
}
\description{
Dataframe with 3 columns containing a list of transcript_id from 
query_gtf, a list of matched ref_transcript_id and the percent coverage between
query and reference transcripts
}
\keyword{datasets}
