% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{query_cds}
\alias{query_cds}
\title{CDS from 4 transcripts entries of the same gene}
\format{
A GRangesList object with 4 elements:
\describe{
  \item{element}{CDS start and end coordinates of each transcript}
  ...
}
}
\source{
\url{http://www.ensembl.org/}
}
\usage{
query_cds
}
\description{
A dataset containing coordinates of CDS from 4 transcripts of
mouse Ptbp1. Transcript names and gene IDs have been modified
}
\keyword{datasets}
