% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{query_exons}
\alias{query_exons}
\title{Exons from 4 transcripts entries of the same gene}
\format{A GRangesList object with 4 elements:
\describe{
  \item{element}{Exon start and end coordinates of each transcript}
  ...
}}
\source{
\url{http://www.ensembl.org/}
}
\usage{
query_exons
}
\description{
A dataset containing coordinates of exons from 4 transcripts of
mouse Ptbp1. Transcript names and gene IDs have been modified
}
\keyword{datasets}
