% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_exons}
\alias{ref_exons}
\title{Exons from 2 reference transcripts entries of the same gene}
\format{A GRangesList object with 2 elements:
\describe{
  \item{element}{Exon start and end coordinates of each transcript}
  ...
}}
\source{
\url{https://www.gencodegenes.org/}
}
\usage{
ref_exons
}
\description{
A dataset containing coordinates of exons from 2 reference transcripts of
mouse Ptbp1.
}
\keyword{datasets}
