% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_gtf}
\alias{ref_gtf}
\title{Imported GTF file containing 2 reference transcript entries of the same gene}
\format{
A GRanges object with 64 ranges and 5 metadata columns:
\describe{
  \item{ranges}{Chromosome, start, end, and strand info of 4 transcripts
  and its exons}
  \item{type}{Entry type; transcript or exon}
  \item{phase}{Phase of open-reading frame}
  \item{gene_id}{Matched gene_id}
  \item{gene_name}{Name of gene}
  \item{transcript_id}{ID given to transcripts}
  ...
}
}
\source{
\url{https://www.gencodegenes.org/}
}
\usage{
ref_gtf
}
\description{
A dataset containing coordinates of transcript and exons from 2 reference
transcripts of mouse Ptbp1.
}
\keyword{datasets}
