% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resizeTranscript.R
\name{resizeTranscript}
\alias{resizeTranscript}
\title{Resize 5' and 3' ends of a transcript GenomicRanges}
\usage{
resizeTranscript(x, start = 0, end = 0)
}
\arguments{
\item{x}{GenomicRanges object containing exon coordinates from a transcript}

\item{start}{Length to append from the start of transcript}

\item{end}{Length to append from the end of transcript}
}
\value{
Appended GenomicRanges object
}
\description{
Resize 5' and 3' ends of a transcript GenomicRanges
}
\examples{
library(GenomicRanges)
gr1 <- GRanges(
  seqnames = "chr1", strand = c("+", "+", "+"),
  ranges = IRanges(
    start = c(1, 500, 1000),
    end = c(100, 600, 1100)
  )
)

resizeTranscript(gr1, 20, 80)
resizeTranscript(gr1, 110, 150)
}
\author{
Fursham Hamid
}
