% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eachfunctions.R
\name{sorteach}
\alias{sorteach}
\title{Internally sort each element of a GenomicRangesList}
\usage{
sorteach(x, ...)
}
\arguments{
\item{x}{GRangesList object}

\item{...}{Comma separated list of unquoted variable names to sort by. Variables are names of metadata columns
found in GRangesList object. Use desc() to sort a variable in descending order. 
Input can be `exonorder` to sort each element in exon order}
}
\value{
Sorted GRangesList object
}
\description{
Internally sort each element of a GenomicRangesList
}
\author{
Fursham Hamid
}
