% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCDS.R
\name{buildCDS}
\alias{buildCDS}
\title{Construct query CDS using reference as guide}
\usage{
buildCDS(query, ref, fasta)
}
\arguments{
\item{query}{GRanges object containing query GTF data.}

\item{ref}{GRanges object containing reference GTF data.}

\item{fasta}{BSgenome or Biostrings object containing genomic sequence}
}
\value{
GRanges object containing query exon entries and newly-constructed CDS
information
}
\description{
buildCDS is designed to construct CDS information on transcripts in query GTF
object. It does this by firstly finding identical query and reference transcript
pairs. If a query is paired to a coding reference transcript, it will be assigned
the reference CDS. For unpaired query transcripts, buildCDS will assign the
upstream-most, annotated ATG codon (if any) as its translation start and search
for an in-frame stop codon.
}
\author{
Fursham Hamid
}
