% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCDS.R
\name{buildCDS}
\alias{buildCDS}
\title{Construct query CDS using reference as guide}
\usage{
buildCDS(query, ref, fasta)
}
\arguments{
\item{query}{GRanges object containing query GTF data.}

\item{ref}{GRanges object containing reference GTF data.}

\item{fasta}{BSgenome or Biostrings object containing genomic sequence}
}
\value{
GRanges object containing query exon entries and newly-constructed cds
information
}
\description{
buildCDS will select the #########

buildCDS will next attempt to construct query CDS information by
firstly deriving its start_codon as guided by the reference transcript. 
This is done in three successive stages:
(1) search transcript for a start_codon as annotated in reference
(2) if above fails, search for an internal ATG codon as annotated in reference
(3) if above fails, align the frame of query to reference

After the start_codon/coding frame have been established, the program
will search for an in-frame stop_codon for each transcript and return its
CDS GRanges entries if successfull.
}
\examples{
library(BSgenome.Mmusculus.UCSC.mm10)
buildCDS(matched_query_gtf, ref_gtf, Mmusculus)

}
\author{
Fursham Hamid
}
