% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputs.R
\name{checkInputs}
\alias{checkInputs}
\title{Title}
\usage{
checkInputs(input, ref, fasta)
}
\arguments{
\item{input}{GRanges object containing imported query GTF data}

\item{ref}{GRanges object containing reference GTF data}

\item{fasta}{BSgenome or Biostrings object containing genomic sequence}
}
\description{
Title
}
