% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chrom_matched_query_gtf}
\alias{chrom_matched_query_gtf}
\title{Chromosome matched version of "query_gtf"}
\format{
A GRanges object with 56 ranges and 3 metadata columns:
\describe{
  \item{ranges}{Chromosome, start, end, and strand info of 4 transcripts
  and its exons}
  \item{type}{Entry type; transcript or exon}
  \item{transcript_id}{ID given to transcripts}
  \item{gene_id}{ID given to gene origin of transcripts}
  ...
}
}
\usage{
data(chrom_matched_query_gtf)
}
\description{
query_gtf data which have been corrected for its seqlevels
}
\keyword{datasets}
