% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateAS.R
\name{compareAS}
\alias{compareAS}
\title{Compare and classify alternative spliced segments between two or more transcripts}
\usage{
compareAS(x, ...)
}
\arguments{
\item{x}{Can be a GRangesList object containing exons for each transcripts (Intralist mode)

Can be a GRanges object containing exons for a transcript (Pair-wise mode). If so,
at least one GRanges object is to be provided in `...` for comparison}

\item{...}{In pair-wise mode, argument is one or more GRanges object containing exons for a
particular transcript to compare with `x`}
}
\value{
GRangesList object containing alternatively-spliced segments for each transcript
in comparison.
}
\description{
Compare and classify alternative spliced segments between two or more transcripts
}
\author{
Fursham Hamid
}
