% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareAS.R
\name{compareAS}
\alias{compareAS}
\title{Compare and classify alternative spliced segments}
\usage{
compareAS(exons, ..., groupings = NULL)
}
\arguments{
\item{exons}{In pair-wise mode: GRanges object containing exons for a particular transcript.

In intra-list mode: GRangesList bject containing exons for each transcripts.
Transcripts will be paired and compared based on its gene family or groups.
In order to do so, object has to contain gene_id attribute. Alternatively, user
may provide a dataframe with a list of transcripts and its groupings as a `groupings`
argument. See `groupings`.}

\item{...}{In pair-wise mode, argument is a GRanges object containing exons for a
particular transcript as a comparison to `exons`}

\item{groupings}{Dataframe describing the groupings of the transcripts in `exons`. Ideally, Transcripts should be
grouped by gene families. Therefore, first column in the dataframe is a list of gene_id or
gene_names and the second column is the names of transcripts in `exons` which fall into
the gene groupings. This argument is optional if gene_id metadata is present in `exons`}
}
\value{
In pairwise mode: a GRangesList object with included and skipped segments in `exons`
as compared to ...

In intralist mode: a dataframe with coordinates and information of alternative segments
between every transcripts in its groupings.
}
\description{
Compare and classify alternative spliced segments
}
