% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{domains.out}
\alias{domains.out}
\title{Example output of predictDomains()}
\format{
A data.frame with 14880 rows and 5 columns:
\describe{
  \item{transcript}{Transcript ID of protein-coding RNAs}
  \item{description}{Name of domain families}
  \item{eval}{E-value score}
  \item{begin}{Start position of domain in protein}
  \item{end}{End position of domain in protein}
  ...
}
}
\usage{
data(domains.out)
}
\description{
Output dataframe from predictDomains() function.
}
\keyword{datasets}
