% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCDSfeature.R
\name{extractCDSfeature}
\alias{extractCDSfeature}
\title{Extract protein features from coding mRNAs}
\usage{
extractCDSfeature(x, fasta, ..., hmmer = "default", signalHsmm = NULL)
}
\arguments{
\item{x}{Can be a GRanges object containing 'CDS' features in GTF format

Can be a GRangesList object containing CDS ranges for each transcript}

\item{fasta}{BSgenome or Biostrings object containing genomic sequence
#' @param ...
Logical conditions to pass to dplyr::filter to subset transcripts for analysis.
Variables are metadata information found in `x` and multiple conditions can be 
provided delimited by comma. Example: transcript_id == "transcript1"}

\item{hmmer}{List of fields to report from hmmer protein domain search. 'default' will
return 'desc', 'evalue' and 'nincluded' columns. See ?bio3d::hmmer for a list
of returned fields. Set argument to NULL to skip hmmer analysis}

\item{signalHsmmm}{List of fields to report from signalHsmmm prediction tool. By default, argument
is set to NULL which skips the analysis. Seting argument to 'default' will
return 'sp_probability' columns. See ?signalHsmm::signalHsmm for a list
of returned fields.}
}
\value{
Dataframe containing protein features for each cds entry
}
\description{
Extract protein features from coding mRNAs
}
\examples{
library(BSgenome.Mmusculus.UCSC.mm10)
extractCDSfeature(query_cds, Mmusculus)
}
