% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importGTF}
\alias{importGTF}
\title{Wrapper to import GTF file}
\usage{
importGTF(con)
}
\arguments{
\item{con}{Path to GTF file}
}
\value{
Imported GenomicRanges object in GTF format
}
\description{
This function is a wrapper to rtracklayer::import() function to conveniently
import GTF file into R as a GenomicRanges object.
}
\author{
Fursham Hamid
}
