% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importGTF}
\alias{importGTF}
\title{Import GTF file into R}
\usage{
importGTF(con)
}
\arguments{
\item{con}{Path to GTF file}
}
\value{
Imported GenomicRanges object in GTF format
}
\description{
This function loads GTF files into R and converts it into


a wrapper to rtracklayer::import() function to conveniently
import GTF file into R as a GenomicRanges object.
}
\examples{
input.gtf <- importGTF(system.file("extdata", "sample.gtf", package = "factR"))
}
\author{
Fursham Hamid
}
