% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchSeqLevels.R
\name{matchSeqLevels}
\alias{matchSeqLevels}
\title{Match seqnames of input GRanges to a reference GRanges}
\usage{
matchSeqLevels(x, to)
}
\arguments{
\item{x}{GRanges object with seqnames to change}

\item{to}{GRanges object from which seqnames is referenced}
}
\value{
Corrected input GRanges
}
\description{
Match seqnames of input GRanges to a reference GRanges
}
\author{
Fursham Hamid
}
