% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{new_query_gtf}
\alias{new_query_gtf}
\title{Query data containing CDS information}
\format{A GRanges object with 105 ranges and 8 metadata columns:
\describe{
  \item{element}{Exon start and end coordinates from 4 transcripts}
  \item{element}{CDS start and end coordinates from 4 transcripts}
  ...
}}
\usage{
new_query_gtf
}
\description{
matched_query_gtf data that has undergone buildCDS function and containing
CDS features
}
\keyword{datasets}
