% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictDomains.R
\name{predictDomains}
\alias{predictDomains}
\title{Predict protein domain families from coding mRNAs}
\usage{
predictDomains(x, fasta, ..., plot = FALSE)
}
\arguments{
\item{x}{Can be a GRanges object containing 'CDS' features in GTF format

Can be a GRangesList object containing CDS ranges for each transcript}

\item{fasta}{BSgenome or Biostrings object containing genomic sequence}

\item{...}{Logical conditions to pass to dplyr::filter to subset transcripts for analysis.
Variables are metadata information found in `x` and multiple conditions can be
provided delimited by comma. Example: transcript_id == "transcript1"}

\item{plot}{Boolean argument to plot out protein domains (Default: FASLE). 
Only first 20 proteins will be plotted}
}
\value{
Dataframe containing protein features for each cds entry
}
\description{
Predict protein domain families from coding mRNAs
}
\author{
Fursham Hamid
}
