% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictNMD.R
\name{predictNMD}
\alias{predictNMD}
\title{Predict sensitivity of mRNA transcripts to NMD}
\usage{
predictNMD(
  exons,
  cds,
  NMD_threshold = 50,
  which = NULL,
  return = c("all", "NMD")
)
}
\arguments{
\item{exons}{GRanges object or GRangesList object containing exons
for each transcript. To search for NMD-inducing features, transcripts have
to be coding and thus contain a cds information of the same transcript name}

\item{cds}{GRanges object or GRangesList object containing coding regions (CDS)
for each transcript. Object type must match `exons` object type. GRangesList 
must have names that match names in exons, else transcript will not be analysed}

\item{NMD_threshold}{Minimum distance of STOP codon to last exon junction (EJ) which triggers NMD.
Default = 50bp}

\item{which}{List containing names of transcripts from `exons` to filter for analysis}

\item{return}{If exons and cds are GRangesList, returns results for all transcripts (defailt) or
only NMD-sensitive (default) transcripts}
}
\value{
List with prediction of NMD sensitivity and statistics:

is_NMD: logical value in prediciting transcript sensitivity to NMD

dist_to_lastEJ: Integer value indicating distance of STOP codon to last EJ
Values are referenced from last EJ, thus a positive value indicates upstream
position of STOP codon while negative value indicates downstream position

num_of_down_EJs: Number of downstream EJs

dist_to_downEJs: Integer value indicating distance of STOP codon to each down
EJs Values are referenced from last EJ, thus a positive value indicates
upstream position of STOP codon while negative value indicates downstream
position together with distances of STOP codon to last EJ
}
\description{
Predict sensitivity of mRNA transcripts to NMD
}
\examples{

### To visualize transcripts
library(wiggleplotr)
plotTranscripts(query_exons, query_cds)

### Examples with GRanges objects
predictNMD(query_exons$transcript1, query_cds$transcript1) # NMD-insensitive
predictNMD(query_exons$transcript3, query_cds$transcript3) # NMD-sensitive

### Examples with GRangesList object
predictNMD(query_exons, query_cds)
predictNMD(query_exons, query_cds, return = "NMD")
predictNMD(query_exons, query_cds, which = c("transcript1", "transcript3"))
}
\author{
Fursham Hamid
}
