% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{query_cds}
\alias{query_cds}
\title{CDS from 4 transcripts entries of the same gene}
\format{
A GRangesList object with 4 elements:
\describe{
  \item{ranges}{Chromosome, start, end, and strand info of 4 transcripts
  and its exons}
  \item{type}{Entry type; transcript or exon}
  \item{transcript_id}{ID given to transcripts}
  \item{phase}{Phase of open-reading frame}
  \item{built_from}{Method by which CDS was built}
  ...
}
}
\usage{
data(query_cds)
}
\description{
A dataset containing coordinates of CDS from 4 transcripts of
mouse Ptbp1. Transcript names and gene IDs have been modified
}
\keyword{datasets}
