% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_exons}
\alias{ref_exons}
\title{Exons from 2 reference transcripts entries of the same gene}
\format{
A GRangesList object with 2 elements:
\describe{
  \item{ranges}{Chromosome, start, end, and strand info of 4 transcripts
  and its exons}
  \item{type}{Entry type; transcript or exon}
  \item{phase}{Phase of open-reading frame}
  \item{gene_id}{Matched gene_id}
  \item{gene_name}{Name of gene}
  \item{transcript_id}{ID given to transcripts}
  ...
}
}
\source{
\url{https://www.gencodegenes.org/}
}
\usage{
data(ref_exons)
}
\description{
A dataset containing coordinates of exons from 2 reference transcripts of
mouse Ptbp1.
}
\keyword{datasets}
