% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchuORFs.R
\name{searchuORFs}
\alias{searchuORFs}
\title{Search coding mRNAs for upstream and overlapping ORFs}
\usage{
searchuORFs(exons, cds, fasta, ORFlength = 21, which = NULL, append = FALSE)
}
\arguments{
\item{exons}{GRanges object or GRangesList object containing exons
for each transcript. To search for uORFs, transcripts have
to be coding and thus contain a cds information of the same transcript name}

\item{cds}{GRanges object or GRangesList object containing coding regions (CDS)
for each transcript. GRangesList must have names that match names in exons,
else exons will not be analysed}

\item{fasta}{BSgenome or Biostrings object containing genomic sequence}

\item{ORFlength}{Minumum length of ORF. Default: 21}

\item{which}{List containing transcript names to filter for analysis}

\item{append}{Logical value to append GRangesList containing uORF ranges to input
exons and cds. Default: FALSE}
}
\value{
List containing exon and cds GRangesList of upstream and overlapping ORFs.
If append is TRUE, function will append newly-found ORFs to input
exons and cds provided
}
\description{
Search coding mRNAs for upstream and overlapping ORFs
}
\examples{

library(BSgenome.Mmusculus.UCSC.mm10)
searchuORFs(query_exons, query_cds, Mmusculus)
searchuORFs(query_exons, query_cds, Mmusculus, append = T)

# To separate exons and cds GRangesList into differebt objects
unpack[query_exons_uORF, query_cds_uORF] = searchuORFs(query_exons, query_cds, Mmusculus)

}
