% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsettranscripts.R
\name{subsetNewTranscripts}
\alias{subsetNewTranscripts}
\title{Subset GTF GRanges object for de novo transcripts}
\usage{
subsetNewTranscripts(query, ref)
}
\arguments{
\item{query}{GRanges object containing query GTF data.}

\item{ref}{GRanges object containing reference GTF data.}
}
\value{
GRanges object containing exon coordinates from transcripts which are only
found in query.
}
\description{
subsetNewTranscripts will compare query and reference GTF GRanges and return
transcripts which are only found in query (de novo).
}
\examples{
subsetNewTranscripts(matched_query_gtf, ref_gtf)
}
