% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewTranscripts.R
\name{viewTranscripts}
\alias{viewTranscripts}
\title{Plot transcripts directly from GTF.}
\usage{
viewTranscripts(x, ..., rescale_introns = F)
}
\arguments{
\item{x}{GRanges object containing transcript annotation in GTF format}

\item{...}{Logical conditions to pass to dplyr::filter to subset transcripts for plotting.
Variables are metadata information found in `x` and multiple conditions can be
provided delimited by comma. Example: gene_name == "Ptbp1"}

\item{rescale_introns}{Specifies if the introns should be scaled to fixed length or not. (default: FALSE)}
}
\value{
ggplot2 object
}
\description{
A wrapper around wiggleplotr's plotTranscripts function. See the documentation for (\code{\link[wiggleplotr]{plotTranscripts}})
for more information.
}
\examples{
## ---------------------------------------------------------------------
## EXAMPLE USING SAMPLE DATASET
## ---------------------------------------------------------------------
viewTranscripts(query_gtf)
viewTranscripts(query_gtf, transcript_id == "transcript1")
viewTranscripts(ref_gtf)

## ---------------------------------------------------------------------
## EXAMPLE USING TRANSCRIPT ANNOTATION
## ---------------------------------------------------------------------
\dontrun{
library(AnnotationHub)

## Retrieve GRCm38 trancript annotation
ah <- AnnotationHub()
GRCm38_gtf <- ah[["AH60127"]]

## Plot transcripts from Ptbp1 gene
viewTranscripts(GRCm38_gtf, gene_name == "Ptbp1")
}

}
\author{
Fursham Hamid
}
